Experiment with some "catch-all" HTML handlers
This adds some handlers to Markdown.Html which unconditionally
return "Ok" or "Err" so that we can experiment with variations of
success/failure

This also demonstrates the behavior when a handler is added to
the CustomHtmlBlockSignupForm.elm to cause any tag failure to
appear as a red "tag error".

I didn't see how to enable rendering the actual name of the
problematic tag with this mechanism so I couldn't include that.